Provision with some seed data
So I tested this and it doesn't matter what you put for the person_id - it will get initialized with 1.

The other interesting thing I learned is that - at least this version of apache xampp - is respecting the FOREIGN KEY directives along with references and on delete cascade.

It will throw an error if the foreign key value doesn not exist, and when the reference is deleted the dependencies will also get deleted if on delete cascade is set - pretty cool