Only 5 login attempts per 24 hours
With this commit a user will be locked out from attempting to login from the same ip address after 5 failed attempts. After 24 hours the failedCount will be reset to 0 giving the user another 5 tries.

I'll not be implementing forgotMyPassword. I think this is already way beyond the intended scope and I've proven that if I really wanted to I could.