Moved counts call out to remove future redundant call
On the edit screan, I'll need to call for the full dataset of both the list, and it's todos. Having the list make an extra call for the todos will be a waist so I'm moving it out so it only happens in this one place. From the edit screen that information can be obtained along with the rest of the todo's information.